requirements.txt
openai
PyMuPDF
streamlit
streamlit-pdf-viewer
python-dotenv
pandas
azure-core
azure-ai-documentintelligence
pillow
python-Levenshtein
pypdf